requirements.txt
numpy
pandas
scikit-learn
matplotlib
librosa
torch
torchaudio
tqdm
umap-learn
sentence-transformers

